+++ src/decorator_trait.jl
@@ -379,24 +379,26 @@ end
 @trait_function is_flat(M::AbstractDecoratorManifold)
 
 # Introduce Deco Trait | automatic foward | fallback
-@trait_function is_point(M::AbstractDecoratorManifold, p, te::Bool = false; kwargs...)
-@trait_function is_point(M::AbstractDecoratorManifold, p, e::Symbol; kwargs...)
+@trait_function is_point(M::AbstractDecoratorManifold, p; kwargs...)
 # Embedded
 function is_point(
     ::TraitList{IsEmbeddedManifold},
     M::AbstractDecoratorManifold,
-    p,
-    te::Bool = false;
+    p;
+    error::Symbol = :none,
     kwargs...,
 )
     # to be safe check_size first
     es = check_size(M, p)
     if es !== nothing
-        te && throw(es)
+        (error === :error) && throw(es)
+        s = "$(typeof(es)) with $(es)"
+        (error === :info) && @info s
+        (error === :warn) && @warn s
         return false
     end
     try
-        pt = is_point(get_embedding(M, p), embed(M, p), te; kwargs...)
+        pt = is_point(get_embedding(M, p), embed(M, p); error = error, kwargs...)
         !pt && return false # no error thrown (deactivated) but returned false -> return false
     catch e
         if e isa DomainError || e isa AbstractManifoldDomainError
@@ -408,28 +410,18 @@ function is_point(
         throw(e) #an error occured that we do not handle ourselves -> rethrow.
     end
     mpe = check_point(M, p; kwargs...)
-    mpe === nothing && return true
-    te && throw(mpe)
-    return false
+    if mpe !== nothing
+        (error === :error) && throw(mpe)
+        s = "$(typeof(mpe)) with $(mpe.val)\n$(mpe.msg)"
+        (error === :info) && @info s
+        (error === :warn) && @warn s
+        return false
+    end
+    return true
 end
 
 # Introduce Deco Trait | automatic foward | fallback
-@trait_function is_vector(
-    M::AbstractDecoratorManifold,
-    p,
-    X,
-    te::Bool = false,
-    cbp = true;
-    kwargs...,
-)
-@trait_function is_vector(
-    M::AbstractDecoratorManifold,
-    p,
-    X,
-    e::Symbol,
-    cbp = true;
-    kwargs...,
-)
+@trait_function is_vector(M::AbstractDecoratorManifold, p, X, cbp::Bool = true; kwargs...)
 # EmbeddedManifold
 # I am not yet sure how to properly document this embedding behaviour here in a docstring.
 function is_vector(
@@ -437,34 +429,41 @@ function is_vector(
     M::AbstractDecoratorManifold,
     p,
     X,
-    te::Bool = false,
-    cbp = true;
+    check_base_point::Bool = true;
+    error::Symbol = :none,
     kwargs...,
 )
     es = check_size(M, p, X)
     if es !== nothing
-        te && throw(es) # error & throw?
+        (error === :error) && throw(es)
+        s = "$(typeof(es)) with $(es)"
+        (error === :info) && @info s
+        (error === :warn) && @warn s
         return false
     end
-    if cbp
-        # check whether p is valid before embedding the tangent vector
-        # throws it te=true
+    if check_base_point
         try
-            ep = is_point(M, p, te; kwargs...)
+            ep = is_point(M, p; error = error, kwargs...)
             !ep && return false
         catch e
             if e isa DomainError || e isa AbstractManifoldDomainError
-                e = ManifoldDomainError(
-                    "$X is not a tangent vector to $p on $M because its bas epoint is not valid point on $M.",
+                ManifoldDomainError(
+                    "$X is not a tangent vector to $p on $M because $p is not a valid point on $p",
                     e,
                 )
             end
             throw(e)
         end
     end
-    # Check vector in embedding
     try
-        tv = is_vector(get_embedding(M, p), embed(M, p), embed(M, p, X), te, cbp; kwargs...)
+        tv = is_vector(
+            get_embedding(M, p),
+            embed(M, p),
+            embed(M, p, X),
+            check_base_point;
+            error = error,
+            kwargs...,
+        )
         !tv && return false # no error thrown (deactivated) but returned false -> return false
     catch e
         if e isa DomainError || e isa AbstractManifoldDomainError
@@ -476,9 +475,12 @@ function is_vector(
         throw(e)
     end
     # Check (additional) local stuff
-    mtve = check_vector(M, p, X; kwargs...)
-    mtve === nothing && return true
-    te && throw(mtve)
+    mXe = check_vector(M, p, X; kwargs...)
+    mXe === nothing && return true
+    (error === :error) && throw(mXe)
+    s = "$(typeof(mXe)) with $(mXe.val)\n$(mXe.msg)"
+    (error === :info) && @info s
+    (error === :warn) && @warn s
     return false
 end