+++ test/domain_errors.jl
@@ -36,38 +36,38 @@ end
     @test !is_point(M, [-1, 1])
     @test !is_point(M, [1, 1, 1]) # checksize fails
     @test_throws DomainError is_point(M, [-1, 1, 1], true) # checksize errors
-    @test_throws DomainError is_point(M, [-1, 1, 1], :error) # checksize errors
+    @test_throws DomainError is_point(M, [-1, 1, 1]; error = :error) # checksize errors
     cs = "DomainError with (3,)\nsize [-1, 1, 1] not (2,)"
-    @test_logs (:info, cs) is_point(M, [-1, 1, 1], :info)
-    @test_logs (:warn, cs) is_point(M, [-1, 1, 1], :warn)
+    @test_logs (:info, cs) is_point(M, [-1, 1, 1]; error = :info)
+    @test_logs (:warn, cs) is_point(M, [-1, 1, 1]; error = :warn)
     @test is_point(M, [1, 1])
-    @test is_point(M, [1, 1], :error)
+    @test is_point(M, [1, 1]; error = :error)
     @test_throws DomainError is_point(M, [-1, 1], true)
-    @test_throws DomainError is_point(M, [-1, 1], :error)
+    @test_throws DomainError is_point(M, [-1, 1]; error = :error)
     ps = "DomainError with [-1, 1]\n<0"
-    @test_logs (:info, ps) is_point(M, [-1, 1], :info)
-    @test_logs (:warn, ps) is_point(M, [-1, 1], :warn)
+    @test_logs (:info, ps) is_point(M, [-1, 1]; error = :info)
+    @test_logs (:warn, ps) is_point(M, [-1, 1]; error = :warn)
 
     @test isa(ManifoldsBase.check_vector(M, [1, 1], [-1, 1]), DomainError)
     @test ManifoldsBase.check_vector(M, [1, 1], [1, 1]) === nothing
     @test !is_vector(M, [1, 1], [-1, 1])
     @test !is_vector(M, [1, 1], [1, 1, 1])
-    @test_throws DomainError is_vector(M, [1, 1], [-1, 1, 1], true)
-    @test_throws DomainError is_vector(M, [1, 1], [-1, 1, 1], :error)
+    @test_throws DomainError is_vector(M, [1, 1], [-1, 1, 1], false, true)
+    @test_throws DomainError is_vector(M, [1, 1], [-1, 1, 1]; error = :error)
     vs = "DomainError with (3,)\nsize [-1, 1, 1] not (2,)"
-    @test_logs (:info, vs) is_vector(M, [1, 1], [-1, 1, 1], :info)
-    @test_logs (:warn, vs) is_vector(M, [1, 1], [-1, 1, 1], :warn)
-    @test !is_vector(M, [1, 1, 1], [1, 1, 1], false, true)
+    @test_logs (:info, vs) is_vector(M, [1, 1], [-1, 1, 1]; error = :info)
+    @test_logs (:warn, vs) is_vector(M, [1, 1], [-1, 1, 1]; error = :warn)
+    @test !is_vector(M, [1, 1, 1], [1, 1, 1], false)
     @test_throws DomainError is_vector(M, [1, 1, 1], [1, 1], true, true)
-    @test_throws DomainError is_vector(M, [1, 1, 1], [1, 1], :error, true)
+    @test_throws DomainError is_vector(M, [1, 1, 1], [1, 1], true; error = :error)
     ps2 = "DomainError with (3,)\nsize [1, 1, 1] not (2,)"
-    @test_logs (:info, ps2) is_vector(M, [1, 1, 1], [1, 1], :info, true)
-    @test_logs (:warn, ps2) is_vector(M, [1, 1, 1], [1, 1], :warn, true)
+    @test_logs (:info, ps2) is_vector(M, [1, 1, 1], [1, 1], true; error = :info)
+    @test_logs (:warn, ps2) is_vector(M, [1, 1, 1], [1, 1], true; error = :warn)
     @test is_vector(M, [1, 1], [1, 1])
-    @test is_vector(M, [1, 1], [1, 1], :none) #default just true/false
-    @test_throws DomainError is_vector(M, [1, 1], [-1, 1], true)
-    @test_throws DomainError is_vector(M, [1, 1], [-1, 1], :error, true)
+    @test is_vector(M, [1, 1], [1, 1]; error = :none)
+    @test_throws DomainError is_vector(M, [1, 1], [-1, 1]; error = :error)
+    @test_throws DomainError is_vector(M, [1, 1], [-1, 1], true; error = :error)
     ps3 = "DomainError with [-1, 1]\n<0"
-    @test_logs (:info, ps3) is_vector(M, [1, 1], [-1, 1], :info, true)
-    @test_logs (:warn, ps3) is_vector(M, [1, 1], [-1, 1], :warn, true)
+    @test_logs (:info, ps3) is_vector(M, [1, 1], [-1, 1], true; error = :info)
+    @test_logs (:warn, ps3) is_vector(M, [1, 1], [-1, 1], true; error = :warn)
 end